feat(libgroup): implement advanced search and filtering capabilities This is a work-in-progress checkpoint focusing on implementing the full search functionality for the LibGroup source template.
**Completed Features:**
- Full search functionality (name, filters, tags, genres)
- Manga Detail & Chapter List viewing
- Chapter Page Reading
- Homepage Listing
- User-configurable settings

**Remaining Tasks:**
- Enhance home screen
- Ensure compatibility with older source

This commit is not ready for merge.

+++ templates/libgroup/src/models/constants.rs
@@ -1,15 +1,17 @@
-use aidoku::alloc::{Vec, string::String};
+use aidoku::alloc::{String, Vec};
 use serde::{Deserialize, Serialize};
 
 #[derive(Default, Deserialize, Serialize, Debug, Clone)]
+#[serde(default)]
 pub struct LibGroupImageServer {
 	pub id: String,
 	pub label: String,
 	pub url: String,
-	pub site_ids: Vec<i32>,
+	pub site_ids: Vec<u8>,
 }
 
 #[derive(Default, Deserialize, Serialize, Debug, Clone)]
+#[serde(default)]
 pub struct LibGroupConstantsData {
 	#[serde(rename = "imageServers")]
 	pub image_servers: Option<Vec<LibGroupImageServer>>,
